requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
pyarrow==14.0.2
pyyaml==6.0.1
prometheus-client==0.19.0
pytest==7.4.3
scipy==1.11.4
numpy==1.26.2
tabulate==0.9.0
matplotlib==3.8.2
